[generate-changelog] Refactor
Just a light reorganization.
